implemented pauliY

+++ include/QubitLayer.h
@@ -9,7 +9,7 @@
 
 constexpr int numQubits = 3;
 
-typedef std::vector<std::complex<float>> qubitStates;
+typedef std::vector<std::complex<double>> qubitStates;
 
 class QubitLayer
 {
